example

+++ src/main/java/pl/pjatk/zjazd5/model/Car.java
@@ -13,5 +13,5 @@ public class Car {
     private final CarClass carClass;
     private boolean isRented;
     //TODO
-    private final Double dailyRate;
+    private final int dailyRate;
 }
